Fix space replacement plugin
The word boundary metacharacter (`\b`) matches (zero-length) on either
the string boundary or a boundary between a word and a non-word
character. Since %, $, etc. are non-word characters, the previous
version of the pattern would fail to match on most reasonable instances.
For instance, it would match on "50 %x" but not on "50 %.".

Fixes #122