Remove unnecessary timeout
This timeout was mistakenly introduced in bfee42b, most likely due to
a semantic error.

`$(document).ready()` expects a function but a naked call to
`.popover()`, not being wrapped in a closure, would evaluate
immediately, returning a jQuery object thus rendering the call to
`.ready()` pointless. The call to `setTimeout()` circumvents this by
calling `.popover()` later, once the DOM is hopefully ready, but the
wrapping call to `.ready()` is still pointless in this case as
`setTimeout()` returns a timeout identifier.