Fix explainer cover (with updated date) on mobile
This commit fixes layout for updated publication dates on mobile. It
makes tags wrap on another line if all the content does not fit on one
line. To this end, it replaces the middot separator with more spacing
(as it is not clear to CSS if it splits to multiple lines).